refactor: 상세페이지에 쓰이는 컴포넌트들을 정리
