+++ components/detail/article/title/index.tsx
@@ -1,0 +1,5 @@
+export default function Title() {
+  return (
+    <div className="flex w-[956px] flex-col items-start gap-40 px-24 py-32"></div>
+  );
+}
