+++ components/detail/tag/index.tsx
@@ -1,0 +1,7 @@
+export default function writerTag() {
+  return (
+    <span className="flex h-24 items-center gap-10 rounded-24 bg-tag-blue-100 px-10 pb-2 pt-3 text-12 font-normal leading-[15.6px] tracking-[-0.6px] text-tag-blue-500 mobile:h-28 mobile:text-14">
+      작성자
+    </span>
+  );
+}
